# cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: X-Admin API
    description: X-Admin API
    contact:
        name: xq
        url: https://github.com/chnxq/xadmin
        email: chnxq@gmail.com
    license:
        name: MIT License
        url: https://github.com/chnxq/xadmin-web/blob/master/LICENSE
    version: "1.0"
paths:
...
    /admin/v1/file/download:
        get:
            tags:
                - FileTransferService
            description: 下载文件
            operationId: FileTransferService_DownloadFile
            parameters:
                - name: fileId
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: storageObject.bucketName
                  in: query
                  schema:
                    type: string
                - name: storageObject.fileDirectory
                  in: query
                  schema:
                    type: string
                - name: storageObject.objectName
                  in: query
                  schema:
                    type: string
                - name: downloadUrl
                  in: query
                  schema:
                    type: string
                - name: rangeStart
                  in: query
                  description: 可选的分段下载范围（闭区间 start，开区间 end；为 0/0 表示全量）
                  schema:
                    type: string
                - name: rangeEnd
                  in: query
                  schema:
                    type: string
                - name: preferPresignedUrl
                  in: query
                  schema:
                    type: boolean
                - name: presignExpireSeconds
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: disposition
                  in: query
                  schema:
                    type: string
                - name: acceptMime
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/DownloadFileResponse'
...
    /admin/v1/files:
        get:
            tags:
                - FileService
            description: 查询文件列表
            operationId: FileService_List
            parameters:
                - name: page
                  in: query
                  description: 当前页码（从1开始，默认1）
                  schema:
                    type: integer
                    format: uint32
                - name: pageSize
                  in: query
                  description: 每页条数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: offset
                  in: query
                  description: 跳过的记录数（从0开始，默认0）
                  schema:
                    type: string
                - name: limit
                  in: query
                  description: 最多返回的记录数（默认10，建议设置上限如100）
                  schema:
                    type: integer
                    format: uint32
                - name: token
                  in: query
                  description: 上一页最后一条记录的游标（如ID/时间戳+ID，首次请求为空）
                  schema:
                    type: string
                - name: noPaging
                  in: query
                  description: 是否不分页，如果为true，则page和pageSize参数无效。
                  schema:
                    type: boolean
                - name: query
                  in: query
                  description: JSON字符串过滤条件，基础语法：{"field1":"val1", "field2___icontains":"val2"}，具体请参见：https://github.com/chnxq/x-crud/tree/main/pagination/filter/README.md
                  schema:
                    type: string
                - name: filter
                  in: query
                  description: Google AIP规范字符串过滤条件
                  schema:
                    type: string
                - name: filterExpr.type
                  in: query
                  description: 过滤表达式类型
                  schema:
                    enum:
                        - EXPR_TYPE_UNSPECIFIED
                        - AND
                        - OR
                    type: string
                    format: enum
                - name: orderBy
                  in: query
                  description: 排序条件
                  schema:
                    type: string
                - name: fieldMask
                  in: query
                  description: 字段掩码，其作用为SELECT中的字段，其语法为使用逗号分隔字段名，例如：id,realName,userName。如果为空则选中所有字段，即SELECT *。
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListFileResponse'
        post:
            tags:
                - FileService
            description: 创建文件
            operationId: FileService_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /admin/v1/files/{id}:
        get:
            tags:
                - FileService
            description: 查询文件详情
            operationId: FileService_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
                - name: viewMask
                  in: query
                  schema:
                    type: string
                    format: field-mask
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/File'
        put:
            tags:
                - FileService
            description: 更新文件
            operationId: FileService_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - FileService
            description: 删除文件
            operationId: FileService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        CreateFileRequest:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/File'
            description: 创建 - 请求
...
        DownloadFileResponse:
            type: object
            properties:
                file:
                    type: string
                    description: 直接返回文件字节（小文件或同步场景）
                    format: bytes
                downloadUrl:
                    type: string
                    description: 预签名 URL（大文件或异步下载）
                sourceFileName:
                    type: string
                    description: 原始文件名，客户端用于保存或显示
                mime:
                    type: string
                    description: MIME 类型，默认 application/octet-stream
                size:
                    type: string
                    description: 文件大小（字节），可用于进度或校验
                checksum:
                    type: string
                    description: 可选，文件校验（如 MD5 或 SHA256）
                storagePath:
                    type: string
                    description: 可选，后端存储位置或 key
                updatedAt:
                    type: string
                    description: 可选，最后修改时间
                    format: date-time
            description: 文件下载响应
...
        File:
            type: object
            properties:
                id:
                    type: integer
                    description: 文件ID
                    format: uint32
                provider:
                    enum:
                        - MINIO
                        - ALIYUN
                        - AWS
                        - AZURE
                        - BAIDU
                        - QINIU
                        - TENCENT
                        - GOOGLE
                        - HUAWEI
                        - LOCAL
                    type: string
                    description: OSS供应商
                    format: enum
                bucketName:
                    type: string
                    description: 存储桶名称
                fileDirectory:
                    example: tenant/1001/user/avatar/
                    type: string
                    description: 文件目录
                fileGuid:
                    type: string
                    description: 文件Guid
                saveFileName:
                    example: f8a3b9c0-1234-5678-90ab-cdef12345678.pdf
                    type: string
                    description: 实际存储文件名（防止在服务器文件系统发生文件冲突）
                fileName:
                    example: 2025年个人所得税申报表.pdf
                    type: string
                    description: 原始文件名
                extension:
                    example: pdf
                    type: string
                    description: 文件扩展名
                size:
                    type: string
                    description: 文件长度，单位：字节
                sizeFormat:
                    example: 2.3 MB
                    type: string
                    description: 格式化后的文件长度字符串
                linkUrl:
                    type: string
                    description: 链接地址
                contentHash:
                    type: string
                    description: 文件内容hash值
                tenantId:
                    type: integer
                    description: 租户ID，0代表系统全局角色
                    format: uint32
                tenantName:
                    type: string
                    description: 租户名称
                createdBy:
                    type: integer
                    description: 创建者用户ID
                    format: uint32
                updatedBy:
                    type: integer
                    description: 更新者用户ID
                    format: uint32
                deletedBy:
                    type: integer
                    description: 删除者用户ID
                    format: uint32
                createdAt:
                    type: string
                    description: 创建时间
                    format: date-time
                updatedAt:
                    type: string
                    description: 更新时间
                    format: date-time
                deletedAt:
                    type: string
                    description: 删除时间
                    format: date-time
            description: 文件
...
        ListFileResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/File'
                total:
                    type: string
            description: 查询列表 - 回应
...
        UpdateFileRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                data:
                    $ref: '#/components/schemas/File'
                updateMask:
                    example: id,realname,username
                    type: string
                    description: 要更新的字段列表
                    format: field-mask
                allowMissing:
                    type: boolean
                    description: 如果设置为true的时候，资源不存在则会新增(插入)，并且在这种情况下`updateMask`字段将会被忽略。
            description: 更新 - 请求
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...