# internal/server/file_manual_http_ext.go
// Code generated by codex.
// generated at: 2026-05-31 18:45:00 CST

package server

import (
	"context"
	"io"
	"net/http"
	"path/filepath"
	"strconv"
	"strings"

	storagev1 "admin/api/gen/storage/v1"
	"admin/internal/data/repo"
	"admin/internal/filex"

	paginationv1 "github.com/chnxq/x-crud/api/gen/pagination/v1"
	"github.com/chnxq/xkitpkg/app"
	httptransport "github.com/chnxq/xkitpkg/transport/http"
)

type generatedDataWithAppCtx interface {
	GetAppCtx() *app.AppCtx
}

type generatedDataWithFileRepo interface {
	FileRepoProvider() repo.FileRepo
}

type fileBrowserPayload struct {
	ID           uint32 `json:"id"`
	FileName     string `json:"fileName"`
	Extension    string `json:"extension"`
	Size         uint64 `json:"size"`
	SizeFormat   string `json:"sizeFormat"`
	BucketName   string `json:"bucketName"`
	FileDirectory string `json:"fileDirectory"`
	ObjectName   string `json:"objectName"`
	TenantID     uint32 `json:"tenantId,omitempty"`
	TenantName   string `json:"tenantName,omitempty"`
	PreviewType  string `json:"previewType"`
	PreviewURL   string `json:"previewUrl"`
	DownloadURL  string `json:"downloadUrl"`
}

func registerManualFileHTTP(srv *httptransport.Server, data GeneratedData) {
	dataWithApp, ok := data.(generatedDataWithAppCtx)
	if !ok || dataWithApp.GetAppCtx() == nil {
		return
	}
	dataWithRepo, ok := data.(generatedDataWithFileRepo)
	if !ok || dataWithRepo.FileRepoProvider() == nil {
		return
	}
	fileTransferService, err := filex.NewTransferManager(dataWithApp.GetAppCtx(), dataWithRepo.FileRepoProvider())
	if err != nil {
		return
	}

	r := srv.Route("/")
	r.POST("/admin/v1/files/upload-multipart", func(ctx httptransport.Context) error {
		httptransport.SetOperation(ctx, "/admin.service.manual.FileBrowser/UploadMultipart")
		h := ctx.Middleware(func(inner context.Context, _ interface{}) (interface{}, error) {
			return handleFileMultipartUpload(inner, ctx, fileTransferService, dataWithRepo.FileRepoProvider())
		})
		out, err := h(ctx, nil)
		if err != nil {
			return err
		}
		return ctx.JSON(http.StatusOK, out)
	})
	r.GET("/admin/v1/files/{id}/preview", func(ctx httptransport.Context) error {
		httptransport.SetOperation(ctx, "/admin.service.manual.FileBrowser/Preview")
		h := ctx.Middleware(func(inner context.Context, _ interface{}) (interface{}, error) {
			return nil, handleFilePreview(inner, ctx, fileTransferService, dataWithRepo.FileRepoProvider())
		})
		_, err := h(ctx, nil)
		return err
	})
	r.GET("/admin/v1/files/{id}/download", func(ctx httptransport.Context) error {
		httptransport.SetOperation(ctx, "/admin.service.manual.FileBrowser/Download")
		h := ctx.Middleware(func(inner context.Context, _ interface{}) (interface{}, error) {
			return nil, handleFileDownload(inner, ctx, fileTransferService, dataWithRepo.FileRepoProvider())
		})
		_, err := h(ctx, nil)
		return err
	})
	r.GET("/admin/v1/files/{id}/browser", func(ctx httptransport.Context) error {
		httptransport.SetOperation(ctx, "/admin.service.manual.FileBrowser/GetBrowserInfo")
		h := ctx.Middleware(func(inner context.Context, _ interface{}) (interface{}, error) {
			return handleFileBrowserInfo(inner, dataWithRepo.FileRepoProvider())
		})
		out, err := h(ctx, nil)
		if err != nil {
			return err
		}
		return ctx.JSON(http.StatusOK, out)
	})
	r.DELETE("/admin/v1/files/{id}/binary", func(ctx httptransport.Context) error {
		httptransport.SetOperation(ctx, "/admin.service.manual.FileBrowser/DeleteBinary")
		h := ctx.Middleware(func(inner context.Context, _ interface{}) (interface{}, error) {
			return nil, handleFileBinaryDelete(inner, fileTransferService)
		})
		_, err := h(ctx, nil)
		if err != nil {
			return err
		}
		return ctx.Result(http.StatusOK, map[string]bool{"success": true})
	})
}

func handleFileMultipartUpload(
	ctx context.Context,
	httpCtx httptransport.Context,
	fileTransferService *filex.TransferManager,
	fileRepo repo.FileRepo,
) (*fileBrowserPayload, error) {
	req := httpCtx.Request()
	if err := req.ParseMultipartForm(32 << 20); err != nil {
		return nil, storagev1.ErrorBadRequest("invalid multipart form")
	}
	file, header, err := req.FormFile("file")
	if err != nil {
		return nil, storagev1.ErrorBadRequest("file field is required")
	}
	defer file.Close()

	content, err := io.ReadAll(file)
	if err != nil {
		return nil, storagev1.ErrorUploadFailed("%v", err)
	}

	uploadReq := &storagev1.UploadFileRequest{
		StorageObject:   &storagev1.StorageObject{},
		Source:          &storagev1.UploadFileRequest_File{File: content},
		SourceFileName:  filex.StringPtr(header.Filename),
		Mime:            filex.StringPtr(header.Header.Get("Content-Type")),
		Size:            filex.Int64Ptr(int64(len(content))),
	}
	if folder := strings.TrimSpace(req.FormValue("folder")); folder != "" {
		uploadReq.StorageObject.FileDirectory = filex.StringPtr(folder)
	}
	if bucket := strings.TrimSpace(req.FormValue("bucketName")); bucket != "" {
		uploadReq.StorageObject.BucketName = filex.StringPtr(bucket)
	}

	uploadResp, err := fileTransferService.UploadFile(ctx, uploadReq)
	if err != nil {
		return nil, err
	}

	listResp, err := fileRepo.List(ctx, paginationRequestNoPaging)
	if err != nil {
		return nil, err
	}
	targetObjectName := uploadResp.GetObjectName()
	for i := len(listResp.GetItems()) - 1; i >= 0; i-- {
		item := listResp.GetItems()[i]
		if item == nil {
			continue
		}
		if item.GetFileName() != header.Filename {
			continue
		}
		if targetObjectName != "" && joinObjectPath(item.GetFileDirectory(), item.GetSaveFileName()) != targetObjectName {
			continue
		}
		return buildFileBrowserPayload(item), nil
	}
	return nil, storagev1.ErrorInternalServerError("uploaded file metadata not found")
}

func handleFilePreview(
	ctx context.Context,
	httpCtx httptransport.Context,
	fileTransferService *filex.TransferManager,
	fileRepo repo.FileRepo,
) error {
	fileID, err := parsePathID(httpCtx)
	if err != nil {
		return err
	}
	record, err := fileRepo.Get(ctx, &storagev1.GetFileRequest{
		QueryBy: &storagev1.GetFileRequest_Id{Id: fileID},
	})
	if err != nil {
		return err
	}
	downloadResp, err := fileTransferService.DownloadFile(ctx, &storagev1.DownloadFileRequest{
		Selector:    &storagev1.DownloadFileRequest_FileId{FileId: fileID},
		Disposition: filex.StringPtr("inline"),
	})
	if err != nil {
		return err
	}

	if url := downloadResp.GetDownloadUrl(); url != "" {
		http.Redirect(httpCtx.Response(), httpCtx.Request(), url, http.StatusTemporaryRedirect)
		return nil
	}
	contentType := detectPreviewContentType(record, downloadResp.GetMime())
	httpCtx.Response().Header().Set("Content-Disposition", buildInlineDisposition(record.GetFileName()))
	httpCtx.Response().Header().Set("Cache-Control", "private, max-age=60")
	return httpCtx.Blob(http.StatusOK, contentType, downloadResp.GetFile())
}

func handleFileDownload(
	ctx context.Context,
	httpCtx httptransport.Context,
	fileTransferService *filex.TransferManager,
	fileRepo repo.FileRepo,
) error {
	fileID, err := parsePathID(httpCtx)
	if err != nil {
		return err
	}
	record, err := fileRepo.Get(ctx, &storagev1.GetFileRequest{
		QueryBy: &storagev1.GetFileRequest_Id{Id: fileID},
	})
	if err != nil {
		return err
	}
	downloadResp, err := fileTransferService.DownloadFile(ctx, &storagev1.DownloadFileRequest{
		Selector:    &storagev1.DownloadFileRequest_FileId{FileId: fileID},
		Disposition: filex.StringPtr("attachment"),
	})
	if err != nil {
		return err
	}
	if url := downloadResp.GetDownloadUrl(); url != "" {
		http.Redirect(httpCtx.Response(), httpCtx.Request(), url, http.StatusTemporaryRedirect)
		return nil
	}
	httpCtx.Response().Header().Set("Content-Disposition", buildAttachmentDisposition(record.GetFileName()))
	return httpCtx.Blob(http.StatusOK, detectPreviewContentType(record, downloadResp.GetMime()), downloadResp.GetFile())
}

func handleFileBrowserInfo(ctx context.Context, fileRepo repo.FileRepo) (*fileBrowserPayload, error) {
	fileID, err := parseContextPathID(ctx)
	if err != nil {
		return nil, err
	}
	record, err := fileRepo.Get(ctx, &storagev1.GetFileRequest{
		QueryBy: &storagev1.GetFileRequest_Id{Id: fileID},
	})
	if err != nil {
		return nil, err
	}
	return buildFileBrowserPayload(record), nil
}

func handleFileBinaryDelete(ctx context.Context, fileTransferService *filex.TransferManager) error {
	fileID, err := parseContextPathID(ctx)
	if err != nil {
		return err
	}
	return fileTransferService.DeleteFile(ctx, fileID)
}

var paginationRequestNoPaging = storagePagingRequestNoPaging()

func storagePagingRequestNoPaging() *paginationv1.PagingRequest {
	value := true
	return &paginationv1.PagingRequest{NoPaging: &value}
}

func parsePathID(ctx httptransport.Context) (uint32, error) {
	value := ctx.Vars().Get("id")
	return parseUint32(value)
}

func parseContextPathID(ctx context.Context) (uint32, error) {
	req, ok := httptransport.RequestFromServerContext(ctx)
	if !ok || req == nil {
		return 0, storagev1.ErrorBadRequest("request context is unavailable")
	}
	parts := strings.Split(strings.Trim(req.URL.Path, "/"), "/")
	if len(parts) < 4 {
		return 0, storagev1.ErrorBadRequest("invalid path")
	}
	return parseUint32(parts[3])
}

func parseUint32(value string) (uint32, error) {
	parsed, err := strconv.ParseUint(strings.TrimSpace(value), 10, 32)
	if err != nil {
		return 0, storagev1.ErrorBadRequest("invalid file id")
	}
	return uint32(parsed), nil
}

func buildFileBrowserPayload(file *storagev1.File) *fileBrowserPayload {
	if file == nil {
		return nil
	}
	objectName := joinObjectPath(file.GetFileDirectory(), file.GetSaveFileName())
	return &fileBrowserPayload{
		ID:            file.GetId(),
		FileName:      file.GetFileName(),
		Extension:     file.GetExtension(),
		Size:          file.GetSize(),
		SizeFormat:    file.GetSizeFormat(),
		BucketName:    file.GetBucketName(),
		FileDirectory: file.GetFileDirectory(),
		ObjectName:    objectName,
		TenantID:      file.GetTenantId(),
		TenantName:    file.GetTenantName(),
		PreviewType:   detectPreviewType(file.GetExtension(), file.GetFileName()),
		PreviewURL:    "/admin/v1/files/" + strconv.FormatUint(uint64(file.GetId()), 10) + "/preview",
		DownloadURL:   "/admin/v1/files/" + strconv.FormatUint(uint64(file.GetId()), 10) + "/download",
	}
}

func detectPreviewType(extension string, fileName string) string {
	ext := strings.TrimPrefix(strings.ToLower(strings.TrimSpace(extension)), ".")
	if ext == "" {
		ext = strings.TrimPrefix(strings.ToLower(filepath.Ext(fileName)), ".")
	}
	switch ext {
	case "jpg", "jpeg", "png", "gif", "webp", "bmp", "svg":
		return "image"
	case "pdf":
		return "pdf"
	case "md", "markdown":
		return "markdown"
	case "drawio", "xml":
		return "drawio"
	case "txt", "json", "yaml", "yml":
		return "text"
	default:
		return "binary"
	}
}

func detectPreviewContentType(file *storagev1.File, fallback string) string {
	contentType := strings.TrimSpace(fallback)
	if contentType != "" {
		return contentType
	}
	switch detectPreviewType(file.GetExtension(), file.GetFileName()) {
	case "image":
		return fileImageMime(file.GetExtension())
	case "pdf":
		return "application/pdf"
	case "markdown":
		return "text/markdown; charset=utf-8"
	case "drawio":
		return "application/xml; charset=utf-8"
	case "text":
		return "text/plain; charset=utf-8"
	default:
		return "application/octet-stream"
	}
}

func fileImageMime(extension string) string {
	switch strings.TrimPrefix(strings.ToLower(extension), ".") {
	case "jpg", "jpeg":
		return "image/jpeg"
	case "png":
		return "image/png"
	case "gif":
		return "image/gif"
	case "webp":
		return "image/webp"
	case "bmp":
		return "image/bmp"
	case "svg":
		return "image/svg+xml"
	default:
		return "application/octet-stream"
	}
}

func buildInlineDisposition(fileName string) string {
	return buildContentDisposition("inline", fileName)
}

func buildAttachmentDisposition(fileName string) string {
	return buildContentDisposition("attachment", fileName)
}

func buildContentDisposition(mode string, fileName string) string {
	name := strings.TrimSpace(fileName)
	if name == "" {
		name = "file"
	}
	return mode + `; filename="` + strings.ReplaceAll(name, `"`, `'`) + `"`
}

func joinObjectPath(directory string, saveFileName string) string {
	dir := strings.Trim(strings.ReplaceAll(directory, "\\", "/"), "/")
	name := strings.Trim(strings.ReplaceAll(saveFileName, "\\", "/"), "/")
	if dir == "" {
		return name
	}
	if name == "" {
		return dir
	}
	return dir + "/" + name
}